Improve appearance of Settings screen. Delete unused screens
