+++ assets/translations/it.json
@@ -103,7 +103,7 @@
 		"restoreBatteryHint": "Si raccomanda di avere almeno il 20% della batteria carica prima di inizializzare il ripristino. Il processo potrebbe durare a lungo, in base alla quantità di dati ",
 		"restoreWifi": "Usare il Wi-Fi.",
 		"restoreWifiHint": "Si raccomanda di utilizzare una connessione Wi-Fi stabile durante il ripristino, per evitare qualsiasi interruzione ed assicurare un recupero veloce e sicuro dei fondi.",
-        "restoreForeground": "Mantienere l'applicazione in primo piano.",
+		"restoreForeground": "Mantienere l'applicazione in primo piano.",
 		"restoreForegroundHint": "Per favore mantienere l'app in primo piano e non terminarla durante il processo di ripristino, così da evitare interruzioni.",
 		"seedMigrationHint": "La mnemonica si applica a tutte le mint, tuttavia ogni mint necessita di un ripristino ad-hoc, dunque è consigliato di annotare frequentemente gli URL delle mint in uso. Se si possiedono fondi ma non ancora una mnemonica, creare un token dalla mint desiredata e riscattarlo dopo aveer generato la nuova mnemonica. Ripetere il processo per ogni mint di cui si vuole fare il backup.",
 		"ecashPayment": "Pagamento ecash",
@@ -244,7 +244,7 @@
 		"imageCacheCleared": "Cache dell'immagine eliminata!",
 		"leaveGithubStar": "Supportaci con una stella su Github",
 		"alreadySpentHint": "Alcune prove che compongono il saldo sono state spese o riscattate altrove. Risolvere l'errore cancellando le prove in 'Opzioni' > 'Gestione mint' > 'Controlla prove'",
-        "useNpub": "Usare NPUB",
+		"useNpub": "Usare NPUB",
 		"autoSwapToDefaulMint": "Scambiare automaticamente con la mint di default",
 		"trustMintOpt": "Affidarsi alla mint",
 		"swapHint": "Questa opzione richiede un pagamento Lightning, un pagamento di commissioni e potrebbe aggiungere una mint sconosciuta alla lista mint.",
@@ -254,14 +254,14 @@
 		"paidInvoice": "{{ count }} richiesta di pagamento pagata con un totale di {{ total }} Sat",
 		"paidInvoices": "{{ count }} richieste di pagamento pagate con un totale di {{ total }} Sat",
 		"checkPayment": "Controllare pagamento",
-		"lnPaymentSpamHint": "Attendere {{remainingSeconds}} secondi per evitare di sovraccaricare la mint.",
+		"lnPaymentSpamHint": "Attendere {{remainingSeconds}} secondi per evitare di sovraccaricare la mint."
 	},
 	"error": {
 		"checkSpendableErr": "Errore di verifica del token",
 		"claimTokenErr": "Errore nel riscatto del token",
 		"header": "Errore",
 		"msg": "Purtroppo si è verificato un errore. Puoi aiutarci a migliorare eNuts facendo uno screenshot e riportandoci il bug.",
-        "reportBug": "Segnalare il bug",
+		"reportBug": "Segnalare il bug",
 		"requestMintErr": "Errore nel recuperare richiesta di pagamento dalla mint.",
 		"stackNA": "Stack di errore non disponibile",
 		"generalMeltingErr": "Non è stato possibile trovare un percorso di pagamento tra la mint e il nodo.",
