+++ src/components/nav/BottomNav.tsx
@@ -29,12 +29,7 @@ export default function BottomNav({ navigation, route }: TBottomNavProps) {
 
 	const isWalletRelatedScreen = route.name === 'dashboard'
 
-	const isSettingsRelatedScreen = route.name === 'Settings' ||
-		route.name === 'General settings' ||
-		route.name === 'Display settings' ||
-		route.name === 'Security settings' ||
-		route.name === 'Privacy settings' ||
-		route.name === 'Contacts settings'
+	const isSettingsRelatedScreen = route.name === 'Settings' || route.name === 'Display settings'
 
 	return (
 		<View
